fix: handle nil kitchen_root config
In rspec tests, `kitchen_root` was `nil`. This just adds a fallback for
when `kitchen_root` is not defined. I believe this would only ever
happen in tests. I confess to trying to set it in the spec but failing
to get it to work with the multiple configs that are used.