Better unique label generation
Now the code will check if another server is already using the
label is wants, and if so it'll pick a different random suffix.

Added retry support if we do end up with a non-unique label
on server create.

Added better error handling on deletes so a 404 doesn't make
it nervous.

Signed-off-by: M. David Bennett <mdavidbennett@syntheticworks.com>